refactor: restructure features for DDD
