+++ features/analytics/tests/report-schedule.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { POST } from '../../app/api/analytics/[orgId]/schedule/route';
+import { POST } from '../../../app/api/analytics/[orgId]/schedule/route';
 import { NextRequest } from 'next/server';
 
 vi.mock('@clerk/nextjs/server', () => ({ auth: () => Promise.resolve({ userId: 'u1' }) }));
