+++ features/vehicles/tests/vehicleActions.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import * as actions from '../lib/actions/vehicleActions';
+import * as actions from '../../../lib/actions/vehicleActions';
 
 vi.mock('next/cache', () => ({ revalidatePath: vi.fn() }));
-vi.mock('../lib/database/db', () => ({
+vi.mock('../../../lib/database/db', () => ({
   __esModule: true,
   default: {
     vehicle: {
@@ -11,8 +11,8 @@ vi.mock('../lib/database/db', () => ({
     }
   }
 }));
-vi.mock('@/lib/errors/handleError', () => ({ handleError: vi.fn() }));
-vi.mock('@/lib/auth/permissions', () => ({ hasPermission: () => true }));
+vi.mock('../../../lib/errors/handleError', () => ({ handleError: vi.fn() }));
+vi.mock('../../../lib/auth/permissions', () => ({ hasPermission: () => true }));
 vi.mock('@/schemas/vehicles', () => ({
   VehicleFormSchema: { safeParse: () => ({ success: true, data: { vin: '1', type: 'tractor', make: 'Make', model: 'Model', year: 2024 } }) }
 }));
